chore: addressed copilot comments
