+++ services/console/src/types.ts
@@ -660,3 +660,28 @@ export interface LocaleOption {
 export interface Locale extends LocaleOption {
     id: UUID
 }
+
+// eslint-disable-next-line @typescript-eslint/no-namespace
+export declare namespace Intl {
+    type Key = 'calendar' | 'collation' | 'currency' | 'numberingSystem' | 'timeZone' | 'unit'
+    function supportedValuesOf(input: Key): string[]
+
+    interface DateTimeFormat {
+        // eslint-disable-next-line @typescript-eslint/method-signature-style
+        format(date?: Date | number): string
+        // eslint-disable-next-line @typescript-eslint/method-signature-style
+        resolvedOptions(): ResolvedDateTimeFormatOptions
+    }
+
+    interface ResolvedDateTimeFormatOptions {
+        locale: string
+        timeZone: string
+        timeZoneName?: string
+    }
+
+    // eslint-disable-next-line no-var
+    var DateTimeFormat: {
+        new(locales?: string | string[]): DateTimeFormat
+        (locales?: string | string[]): DateTimeFormat
+    }
+}
