+++ services/console/src/views/project/ProjectForm.tsx
@@ -1,37 +1,12 @@
 import api from '../../api'
 import TextInput from '../../ui/form/TextInput'
-import { Project } from '../../types'
+import { Project, Intl } from '../../types'
 import FormWrapper from '../../ui/form/FormWrapper'
 import { SingleSelect } from '../../ui/form/SingleSelect'
 import SwitchField from '../../ui/form/SwitchField'
 import Heading from '../../ui/Heading'
 import { LocaleTextField } from '../settings/Locales'
 import { useTranslation } from 'react-i18next'
-
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export declare namespace Intl {
-    type Key = 'calendar' | 'collation' | 'currency' | 'numberingSystem' | 'timeZone' | 'unit'
-    function supportedValuesOf(input: Key): string[]
-
-    interface DateTimeFormat {
-        // eslint-disable-next-line @typescript-eslint/method-signature-style
-        format(date?: Date | number): string
-        // eslint-disable-next-line @typescript-eslint/method-signature-style
-        resolvedOptions(): ResolvedDateTimeFormatOptions
-    }
-
-    interface ResolvedDateTimeFormatOptions {
-        locale: string
-        timeZone: string
-        timeZoneName?: string
-    }
-
-    // eslint-disable-next-line no-var
-    var DateTimeFormat: {
-        new(locales?: string | string[]): DateTimeFormat
-        (locales?: string | string[]): DateTimeFormat
-    }
-}
 
 interface ProjectFormProps {
     project?: Project
